pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mydomainlist"
version = "0.0.1"
description = "A desktop gui program to create your own domain list."
authors = [
    { name = "ahmetcanisik", email = "can.isik.business@gmail.com" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.8"
keywords = ["domain", "gui", "markdown", "whois"]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]
dependencies = [
    "pillow",
    "customtkinter",
    "CTkMessagebox",
    "python-whois",
    "CTkToolTip",
    "rich"
]

[project.urls]
Homepage = "https://github.com/ahmetcanisik/mydomainlist.git"

[project.scripts]
mydomainlist = "app:main" 